+++ src/main/java/dev/paedar/aoc/lvl17/opcode/BDV.java
@@ -7,7 +7,7 @@ public final class BDV extends Divider {
     public static final BDV INSTANCE = new BDV();
 
     @Override
-    void storeResult(int result, Computer computer) {
+    void storeResult(long result, Computer computer) {
         computer.setRegisterB(result);
     }
