+++ src/main/java/dev/paedar/aoc/lvl17/opcode/BST.java
@@ -7,7 +7,7 @@ public final class BST extends ComboConsumer {
     public static final BST INSTANCE = new BST();
 
     @Override
-    void execute(int compoOperand, Computer computer) {
+    void execute(long compoOperand, Computer computer) {
         computer.setRegisterB(OpCode.mod8(compoOperand));
     }
