+++ src/main/java/dev/paedar/aoc/lvl17/opcode/CDV.java
@@ -7,7 +7,7 @@ public final class CDV extends Divider {
     public static final CDV INSTANCE = new CDV();
 
     @Override
-    void storeResult(int result, Computer computer) {
+    void storeResult(long result, Computer computer) {
         computer.setRegisterC(result);
     }
