+++ src/main/java/dev/paedar/aoc/lvl17/opcode/OUT.java
@@ -7,8 +7,8 @@ public final class OUT extends ComboConsumer {
     public static final OUT INSTANCE = new OUT();
 
     @Override
-    void execute(int compoOperand, Computer computer) {
-        computer.output(OpCode.mod8(compoOperand));
+    void execute(long compoOperand, Computer computer) {
+        computer.output((int) OpCode.mod8(compoOperand));
     }
 
 }
